Broke out packing builder and utils into separate packages (#1565)
* Add packing packages for builder and utils

* Add packing packages for builder and utils
