+++ heron/packing/tests/java/BUILD
@@ -70,7 +70,8 @@ java_test(
     srcs = glob(
        ["**/roundrobin/ResourceCompliantRRPackingTest.java"]
     ),
-    deps = roundrobin_deps_files,
+    deps = roundrobin_deps_files +
+           ["//heron/packing/src/java:utils"],
     size = "small",
 )
 
@@ -79,7 +80,8 @@ java_test(
     srcs = glob(
         ["**/binpacking/FirstFitDecreasingPackingTest.java"]
     ),
-    deps = binpacking_deps_files,
+    deps = binpacking_deps_files +
+           ["//heron/packing/src/java:utils"],
     size = "small",
 )
 
@@ -88,7 +90,8 @@ java_test(
     srcs = glob(
         ["**/PackingUtilsTest.java"]
     ),
-    deps = packing_utils_deps_files,
+    deps = packing_utils_deps_files +
+           ["//heron/packing/src/java:utils"],
     size = "small",
 )
 
@@ -97,7 +100,8 @@ java_test(
     srcs = glob(
         ["**/PackingPlanBuilderTest.java"]
     ),
-    deps = packing_utils_deps_files,
+    deps = packing_utils_deps_files +
+           ["//heron/packing/src/java:builder"],
     size = "small",
 )
