+++ heron/packing/tests/java/com/twitter/heron/packing/utils/PackingUtilsTest.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.twitter.heron.packing;
+package com.twitter.heron.packing.utils;
 
 import java.util.HashMap;
 import java.util.Map;
